requirements.txt
# ic_routing_board_generation
chex<0.1.6
hydra-core==1.3.2
jax==0.4.8
jumanji==0.2.2
matplotlib==3.7.1
numpy>=1.19.5
omegaconf==2.3.0
optax==0.1.5
pygame==2.4.0
pytest==7.3.1

# Visualising boards
opencv-python
opencv-python-headless

# agent_training
tqdm==4.65.0

# jumanji
jumanji
dm-haiku
dm_env==1.6
dm_tree==0.1.8
gym==0.26.2
ipython==8.13.1
neptune==1.1.1
pytest_mock==3.10.0
rlax==0.1.5
scipy==1.10.1
tensorboardX==2.6
typing_extensions==4.5.0
